+++ lib/server/server.ts
@@ -16,7 +16,7 @@ import OktaAuthBase from '../OktaAuthBase';
 import fetchRequest from '../fetch/fetchRequest';
 import { getUserAgent } from '../builderUtil';
 import serverStorage from './serverStorage';
-import { isSigninOptions } from '../types';
+import { isSignInWithCredentialsOptions } from '../types';
 import { AuthSdkError } from '../errors';
 const PACKAGE_JSON = require('../../package.json');
 
@@ -34,7 +34,7 @@ export default class OktaAuthNode extends OktaAuthBase {
   }
 
   signIn(opts) {
-    if (!isSigninOptions(opts)) {
+    if (!isSignInWithCredentialsOptions(opts)) {
       throw new AuthSdkError('Invalid signinOptions are provided.');
     }
     return super.signIn(opts);